Extend allow/block/reject-device commands to accept a rule as a single argument  - Fixes an issue when the rule contains only the target  - Refactoring (usbguard-{allow|block|reject}-device.cpp files)
